refactor nifs to vanilafs
