+++ src/sps.rs
@@ -19,11 +19,12 @@ pub enum Error {
 }
 
 /// This trait verifies whether the instance is faithly generated by a Special soundness protocol (sps)
-pub trait SpecialSoundnessVerifier<C: CurveAffine, RO: ROTrait<C>> {
+/// Reference: section 3.1 of [protostar](https://eprint.iacr.org/2023/620)
+pub trait SpecialSoundnessVerifier<C: CurveAffine, RO: ROTrait<C::Base>> {
     fn sps_verify(&self, ro_nark: &mut RO) -> Result<(), Error>;
 }
 
-impl<C: CurveAffine, RO: ROTrait<C>> SpecialSoundnessVerifier<C, RO> for PlonkInstance<C> {
+impl<C: CurveAffine, RO: ROTrait<C::Base>> SpecialSoundnessVerifier<C, RO> for PlonkInstance<C> {
     fn sps_verify(&self, ro_nark: &mut RO) -> Result<(), Error> {
         let num_challenges = self.challenges.len();
         if num_challenges == 0 {
@@ -31,11 +32,11 @@ impl<C: CurveAffine, RO: ROTrait<C>> SpecialSoundnessVerifier<C, RO> for PlonkIn
         }
 
         self.instance.iter().for_each(|inst| {
-            ro_nark.absorb_base(fe_to_fe(inst).unwrap());
+            ro_nark.absorb_field(fe_to_fe(inst).unwrap());
         });
         for i in 0..num_challenges {
             ro_nark.absorb_point(&self.W_commitments[i]);
-            let r = ro_nark.squeeze(NUM_CHALLENGE_BITS);
+            let r = ro_nark.squeeze::<C>(NUM_CHALLENGE_BITS);
             if r != self.challenges[i] {
                 return Err(Error::ChallengeNotMatch { challenge_index: i });
             }
